Modify the row and column display to what can be seen from the back
When programming the MMX, one is looking at the back of the machine.
Here the lanes with a higher number are on the left, and bass + drums are similarly on the left.

Also, at the front each track has a unique number: vibraphone 1, 2, 3, ... to 22,
the other channels have a track 1 and 2 instead of left or right.

Finally, change the row numbering to have a plate id prefix and count from the bottom
(the registrators are at the top and will read the program as it passes towards
the programmer in the back). The number is now 1-16 for each quarter note beat
and restarts on a plate edge.