Make sure instrument columns have correct order
Apparently the "index=..." property defines the order of the columns
because React seems to sort on it.

Make sure it starts counting from the bass channels and increases for
the vibraphone bars.

Also, delimit index and instrument name, as it meant that
vibraphonebar1-18 and vibraphonebar11-8 were not unique if the dash is
not there.