Add radio button to enable placing pins by hand
The default is to only allow clicking on existing pins (from a loaded
program or imported MIDI) to move them to the other channel.

This was a hidden setting before but it might be better to expose it,
though the radio group might need to be moved to its own dialog.